+++ textrun-npm/src/actions/script-call.ts
@@ -1,5 +1,5 @@
-import * as color from "colorette"
 import { promises as fs } from "fs"
+import { styleText } from "node:util"
 import * as textRunner from "text-runner-engine"
 
 import { startsWithNpmRun } from "../helpers/starts-with-npm-run.js"
@@ -15,7 +15,7 @@ export async function scriptCall(action: textRunner.actions.Args): Promise<void>
     throw new Error(`Does not start with "npm run": ${call}`)
   }
   const want = trimNpmRun(call)
-  action.name(`verify call of npm package script ${color.cyan(want)}`)
+  action.name(`verify call of npm package script ${styleText("cyan", want)}`)
   const packageJsonPath = action.configuration.sourceDir.joinStr("package.json")
   const pkgText = await fs.readFile(packageJsonPath, "utf-8")
   const pkgData: PackageJson = JSON.parse(pkgText)