+++ textrun-npm/src/actions/script-name.ts
@@ -1,5 +1,5 @@
-import * as color from "colorette"
 import { promises as fs } from "fs"
+import { styleText } from "node:util"
 import * as textRunner from "text-runner-engine"
 
 import { PackageJson } from "./package-json.js"
@@ -9,7 +9,7 @@ export async function scriptName(action: textRunner.actions.Args): Promise<void>
   if (want === "") {
     throw new Error("No script name specified")
   }
-  action.name(`npm package has script ${color.cyan(want)}`)
+  action.name(`npm package has script ${styleText("cyan", want)}`)
   const packageJsonPath = action.configuration.sourceDir.joinStr("package.json")
   const pkgText = await fs.readFile(packageJsonPath, "utf-8")
   const pkgData: PackageJson = JSON.parse(pkgText)