+++ textrun-workspace/src/actions/working-dir.ts
@@ -1,10 +1,10 @@
-import * as color from "colorette"
+import { styleText } from "node:util"
 import * as textRunner from "text-runner-engine"
 
 /** The "cd" action changes the current working directory to the one given in the hyperlink or code block. */
 export function workingDir(action: textRunner.actions.Args): void {
   const directory = action.region.text()
-  action.name(`changing into the ${color.cyan(directory)} directory`)
+  action.name(`changing into the ${styleText("cyan", directory)} directory`)
   const fullPath = action.configuration.workspace.joinStr(directory)
   action.log(`cd ${fullPath}`)
   try {